+++ INFO450SaveMore/Checking.h
@@ -8,7 +8,7 @@ private:
 	const double PURCHASE_CHECK_PRICE = 15.00;
 
 public:
-	Checking(int number, int balance); //Constructor
+	Checking(int number, double balance); //Constructor
 	int withdrawFunds(double withdrawAmount); // Needs to be overridden. 
 	int orderChecks(); 	// for checking account
 };
